+++ src/modules/Intonation/duffint.cc
@@ -84,7 +84,7 @@ LISP FT_Int_Targets_Relation_Utt(LISP utt, LISP relname)
     EST_Track *pm = 0;
     LISP params;
     float start,end;
-    int n_frames;
+    ssize_t n_frames;
     
     *cdebug << "Intonation duff module\n";
 
@@ -110,7 +110,7 @@ LISP FT_Int_Targets_Relation_Utt(LISP utt, LISP relname)
     float m = (end-start) /end_time;
     float c = start;
 
-    for (int i = 0; i < n_frames; ++ i)
+    for (ssize_t i = 0; i < n_frames; ++ i)
 	f0->a(i) = (m * ((float) i) * 0.01) + c;
 
     u->create_relation("f0");
